Revert "[SYCL][NFC] Refactor [[intel::reqd_sub_group_size()]] attribute tests (#6243)"
This reverts commit 4071659a3fa58551d8803f6c3ffc191ffb3fda4b.
